# pb-laravel/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "APIs For Thrift Store",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/job-postings/{id}": {
            "get": {
                "tags": [
                    "JobPostings"
                ],
                "summary": "Find job posting by ID",
                "description": "Returns a single job posting",
                "operationId": "findJobPostingById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of job posting to return",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "title": {
                                            "type": "string",
                                            "example": "Software Engineer"
                                        },
                                        "description": {
                                            "type": "string",
                                            "example": "Job description here..."
                                        },
                                        "location": {
                                            "type": "string",
                                            "example": "Istanbul"
                                        },
                                        "posted_by": {
                                            "type": "integer",
                                            "example": 2
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2025-08-06T12:34:56Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2025-08-06T12:34:56Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Job posting not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "JobPostings"
                ],
                "summary": "Update an existing job posting",
                "description": "Updates the details of an existing job posting",
                "operationId": "updateJobPosting",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the job posting to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Senior Software Engineer"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Updated job description..."
                                    },
                                    "location": {
                                        "type": "string",
                                        "example": "Ankara"
                                    },
                                    "posted_by": {
                                        "type": "integer",
                                        "example": 3
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Job posting updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "title": {
                                                    "type": "string",
                                                    "example": "Senior Software Engineer"
                                                },
                                                "description": {
                                                    "type": "string",
                                                    "example": "Updated job description..."
                                                },
                                                "location": {
                                                    "type": "string",
                                                    "example": "Ankara"
                                                },
                                                "posted_by": {
                                                    "type": "integer",
                                                    "example": 3
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Job posting not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "JobPostings"
                ],
                "summary": "Delete a job posting",
                "description": "Deletes a job posting by ID",
                "operationId": "deleteJobPosting",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the job posting to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Job posting deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Job posting deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Job posting not found"
                    }
                }
            }
        },
        "/api/job-postings": {
            "post": {
                "tags": [
                    "JobPostings"
                ],
                "summary": "Create a new job posting",
                "description": "Creates a new job posting in the system",
                "operationId": "createJobPosting",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "description",
                                    "location",
                                    "posted_by"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Software Engineer"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Job description here..."
                                    },
                                    "location": {
                                        "type": "string",
                                        "example": "Istanbul"
                                    },
                                    "posted_by": {
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Job posting created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "title": {
                                                    "type": "string",
                                                    "example": "Software Engineer"
                                                },
                                                "description": {
                                                    "type": "string",
                                                    "example": "Job description here..."
                                                },
                                                "location": {
                                                    "type": "string",
                                                    "example": "Istanbul"
                                                },
                                                "posted_by": {
                                                    "type": "integer",
                                                    "example": 2
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}